requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==3.2.17
django-environ==0.11.2
dotmap==1.3.30
exceptiongroup==1.2.0
frozenlist==1.4.1
func-timeout==4.3.5
gunicorn==21.2.0
h11==0.14.0
idna==3.6
multidict==6.0.5
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==23.2
py-cord==2.4.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.10.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
